solana: Refactor out`transfer_[ownership | token_authority]` into separate files
